sort the utils functions
